Split build cache into layered modules
