+++ crates/unpack_core/src/build_cache.rs
@@ -1,39 +1,50 @@
+//! Compiler-owned Build Cache composition root.
+//! It wires typed Cache Facades to Cache Layers and owns preparation and publication lifecycle.
+
 use std::{
-    any::Any,
-    collections::{BTreeSet, HashMap},
+    collections::BTreeSet,
     fmt, fs, io,
     marker::PhantomData,
-    path::{Path, PathBuf},
+    path::Path,
     sync::{
-        Arc, Mutex, OnceLock,
+        Arc, Mutex,
         atomic::{AtomicBool, AtomicU64, Ordering},
     },
-    time::{Duration, Instant, SystemTime, UNIX_EPOCH},
 };
 
 use crate::{
     InfrastructureLogEvent, InfrastructureLogLevel, ModuleIdentity, SnapshotOptions,
     SnapshotStrategy, UnpackResolver,
     code_generation_record::CodeGenerationRecord,
-    module::BuiltModuleContent,
-    pack_file::{
-        AccessStamp, AssetRenderRecordCodec, AssetRenderRecordDto, CodeGenerationRecordCodec,
-        CodeGenerationRecordDto, CodecRegistry, DEFAULT_MAX_AGE, ModuleBuildRecordCodec,
-        ModuleBuildRecordDto, PackFile, PackFileAddress, PackFileCompression, PackFileETag,
-        PackFileGuardDto, PackFileOpenOptions, PackFilePublicationOptions,
-        PackFileRestore as PackFileRecordRestore, PackFileRetention, PackFileWriteBatch,
-        PublicationBase, ResolveRecordCodec, ResolveRecordDto, SnapshotDto,
-    },
-    parser::ParsedModule,
+    pack_file::{AccessStamp, PackFileGuardDto, SnapshotDto},
     rendered_source::RenderedSource,
     snapshot::{FileSystemInfo, Snapshot, SnapshotCache},
 };
 
-const RESOLVE_CACHE_NAMESPACE: CacheNamespace = CacheNamespace::new("unpack/resolve");
-const MODULE_BUILD_CACHE_NAMESPACE: CacheNamespace = CacheNamespace::new("unpack/module-build");
-const CODE_GENERATION_CACHE_NAMESPACE: CacheNamespace =
-    CacheNamespace::new("unpack/code-generation");
-const ASSET_RENDER_CACHE_NAMESPACE: CacheNamespace = CacheNamespace::new("unpack/asset-render");
+mod cache;
+mod facade;
+mod memory;
+mod options;
+mod persistent;
+mod records;
+#[cfg(test)]
+mod tests;
+
+pub use options::{BuildDependency, CacheCompression, CacheKind, CacheOptions};
+
+use cache::Cache;
+pub(crate) use cache::{CacheItemFamily, CacheItemWork, CacheWorkCounters};
+use facade::{
+    ASSET_RENDER_CACHE_NAMESPACE, CODE_GENERATION_CACHE_NAMESPACE, MODULE_BUILD_CACHE_NAMESPACE,
+    RESOLVE_CACHE_NAMESPACE,
+};
+pub(crate) use facade::{
+    CacheETag, CacheFacade, CacheIdentifier, CacheKey, CacheNamespace, ModuleBuildCache,
+    NormalModuleFactoryCache,
+};
+use persistent::PersistentCachePreparation;
+pub(crate) use records::{ModuleBuildRecord, ResolveRecord, ResolveRequest};
+
 const CACHE_PROFILE_LOGGER: &str = "unpack.Cache.Profile";
 const CACHE_WRITER_LOGGER: &str = "unpack.Cache.Writer";
 
@@ -94,235 +105,6 @@ pub(crate) struct BuildCache {
     diagnostics: Arc<CacheDiagnostics>,
     clock: Arc<dyn CacheClock>,
     inner: Arc<BuildCacheInner>,
-}
-
-#[derive(Debug, Clone)]
-pub(crate) struct CacheFacade<K, V> {
-    build_cache: BuildCache,
-    namespace: CacheNamespace,
-    family: CacheItemFamily,
-    marker: PhantomData<fn(K) -> V>,
-}
-
-pub(crate) type NormalModuleFactoryCache = CacheFacade<ResolveRequest, ResolveRecord>;
-pub(crate) type ModuleBuildCache = CacheFacade<ModuleIdentity, ModuleBuildRecord>;
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-pub(crate) struct CacheNamespace(&'static str);
-
-impl CacheNamespace {
-    pub(crate) const fn new(value: &'static str) -> Self {
-        Self(value)
-    }
-
-    const fn as_str(self) -> &'static str {
-        self.0
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Hash)]
-pub(crate) struct CacheIdentifier(Vec<u8>);
-
-impl CacheIdentifier {
-    #[allow(dead_code)]
-    pub(crate) fn new(value: impl AsRef<[u8]>) -> Self {
-        Self(value.as_ref().to_vec())
-    }
-
-    pub(crate) fn from_parts(parts: impl IntoIterator<Item = Vec<u8>>) -> Self {
-        let mut bytes = Vec::new();
-        for part in parts {
-            bytes.extend_from_slice(&(part.len() as u64).to_le_bytes());
-            bytes.extend_from_slice(&part);
-        }
-        Self(bytes)
-    }
-
-    pub(crate) fn as_bytes(&self) -> &[u8] {
-        &self.0
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub(crate) struct CacheETag(Vec<u8>);
-
-impl CacheETag {
-    #[allow(dead_code)]
-    pub(crate) fn new(value: impl AsRef<[u8]>) -> Self {
-        Self(value.as_ref().to_vec())
-    }
-
-    fn as_bytes(&self) -> &[u8] {
-        &self.0
-    }
-}
-
-pub(crate) trait CacheKey: Clone + Send + Sync + 'static {
-    fn cache_identifier(&self) -> CacheIdentifier;
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub(crate) enum CacheItemFamily {
-    Resolve,
-    ModuleBuild,
-    CodeGeneration,
-    AssetRender,
-}
-
-impl CacheItemFamily {
-    const fn index(self) -> usize {
-        match self {
-            Self::Resolve => 0,
-            Self::ModuleBuild => 1,
-            Self::CodeGeneration => 2,
-            Self::AssetRender => 3,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, Default, PartialEq, Eq)]
-pub(crate) struct CacheItemWork {
-    pub hits: usize,
-    pub misses: usize,
-    pub stores: usize,
-    pub restores: usize,
-    pub evictions: usize,
-}
-
-#[derive(Debug, Clone, Copy, Default, PartialEq, Eq)]
-pub(crate) struct CacheWorkCounters {
-    by_family: [CacheItemWork; 4],
-}
-
-impl CacheWorkCounters {
-    pub(crate) fn for_family(self, family: CacheItemFamily) -> CacheItemWork {
-        self.by_family[family.index()]
-    }
-
-    fn for_family_mut(&mut self, family: CacheItemFamily) -> &mut CacheItemWork {
-        &mut self.by_family[family.index()]
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct CacheOptions {
-    pub kind: CacheKind,
-    pub cache_directory: Option<PathBuf>,
-    pub cache_location: Option<PathBuf>,
-    pub name: Option<String>,
-    pub version: Option<String>,
-    pub build_dependencies: Vec<BuildDependency>,
-    pub automatic_build_dependencies: Vec<PathBuf>,
-    pub max_age: Duration,
-    pub max_memory_generations: Option<u64>,
-    pub compression: CacheCompression,
-    pub allow_collecting_memory: bool,
-    pub idle_timeout: Option<u32>,
-    pub idle_timeout_for_initial_store: Option<u32>,
-    pub idle_timeout_after_large_changes: Option<u32>,
-    pub profile: bool,
-    pub readonly: bool,
-}
-
-impl Default for CacheOptions {
-    fn default() -> Self {
-        Self::memory()
-    }
-}
-
-impl CacheOptions {
-    pub fn disabled() -> Self {
-        Self {
-            kind: CacheKind::Disabled,
-            cache_directory: None,
-            cache_location: None,
-            name: None,
-            version: None,
-            build_dependencies: Vec::new(),
-            automatic_build_dependencies: Vec::new(),
-            max_age: DEFAULT_MAX_AGE,
-            compression: CacheCompression::None,
-            allow_collecting_memory: false,
-            max_memory_generations: None,
-            idle_timeout: None,
-            idle_timeout_for_initial_store: None,
-            idle_timeout_after_large_changes: None,
-            profile: false,
-            readonly: false,
-        }
-    }
-
-    pub fn memory() -> Self {
-        Self {
-            kind: CacheKind::Memory,
-            cache_directory: None,
-            cache_location: None,
-            name: None,
-            version: None,
-            build_dependencies: Vec::new(),
-            automatic_build_dependencies: Vec::new(),
-            max_age: DEFAULT_MAX_AGE,
-            compression: CacheCompression::None,
-            allow_collecting_memory: false,
-            max_memory_generations: None,
-            idle_timeout: None,
-            idle_timeout_for_initial_store: None,
-            idle_timeout_after_large_changes: None,
-            profile: false,
-            readonly: false,
-        }
-    }
-
-    pub fn filesystem() -> Self {
-        Self {
-            kind: CacheKind::Filesystem,
-            cache_directory: None,
-            cache_location: None,
-            name: None,
-            version: None,
-            build_dependencies: Vec::new(),
-            automatic_build_dependencies: Vec::new(),
-            max_age: DEFAULT_MAX_AGE,
-            compression: CacheCompression::None,
-            allow_collecting_memory: false,
-            max_memory_generations: None,
-            idle_timeout: Some(60_000),
-            idle_timeout_for_initial_store: Some(5_000),
-            idle_timeout_after_large_changes: Some(1_000),
-            profile: false,
-            readonly: false,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum CacheKind {
-    Disabled,
-    Memory,
-    Filesystem,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum CacheCompression {
-    None,
-    Gzip,
-    Brotli,
-}
-
-impl From<CacheCompression> for PackFileCompression {
-    fn from(compression: CacheCompression) -> Self {
-        match compression {
-            CacheCompression::None => Self::None,
-            CacheCompression::Gzip => Self::Gzip,
-            CacheCompression::Brotli => Self::Brotli,
-        }
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct BuildDependency {
-    pub name: String,
-    pub requests: Vec<String>,
 }
 
 #[derive(Debug)]
@@ -439,1359 +221,6 @@ impl BuildCacheInner {
             .expect("cache generation update should always succeed");
     }
 }
-
-#[derive(Debug, Clone, PartialEq, Eq, Hash)]
-struct CacheAddress {
-    namespace: CacheNamespace,
-    identifier: CacheIdentifier,
-}
-
-impl CacheAddress {
-    fn to_pack_file_address(&self) -> PackFileAddress {
-        PackFileAddress::new(self.namespace.as_str(), self.identifier.as_bytes())
-    }
-}
-
-impl CacheETag {
-    fn to_pack_file_etag(&self) -> PackFileETag {
-        PackFileETag::new(self.as_bytes())
-    }
-}
-
-#[derive(Clone)]
-struct CacheEntry {
-    family: CacheItemFamily,
-    etag: Option<CacheETag>,
-    value: Arc<dyn Any + Send + Sync>,
-}
-
-impl CacheEntry {
-    fn new<V>(family: CacheItemFamily, etag: Option<CacheETag>, value: V) -> Self
-    where
-        V: Send + Sync + 'static,
-    {
-        Self {
-            family,
-            etag,
-            value: Arc::new(value),
-        }
-    }
-
-    fn value<V: Send + Sync + 'static>(&self) -> Option<Arc<V>> {
-        Arc::clone(&self.value).downcast::<V>().ok()
-    }
-}
-
-impl fmt::Debug for CacheEntry {
-    fn fmt(&self, formatter: &mut fmt::Formatter<'_>) -> fmt::Result {
-        formatter
-            .debug_struct("CacheEntry")
-            .field("family", &self.family)
-            .field("etag", &self.etag)
-            .finish_non_exhaustive()
-    }
-}
-
-#[derive(Debug)]
-enum CacheLayerLookup {
-    Hit(CacheEntry),
-    Deferred(PersistentRestore),
-    Miss,
-}
-
-enum PreparedPersistentRecord {
-    Resolve(PackFileRecordRestore<ResolveRecordDto>),
-    ModuleBuild(PackFileRecordRestore<ModuleBuildRecordDto>),
-    CodeGeneration(PackFileRecordRestore<CodeGenerationRecordDto>),
-    AssetRender(PackFileRecordRestore<AssetRenderRecordDto>),
-}
-
-#[derive(Debug, Clone)]
-struct PersistentRestore {
-    pack_file: Arc<Mutex<PackFile>>,
-    reader_generation: u64,
-    address: PackFileAddress,
-    etag: Option<PackFileETag>,
-    cache_etag: Option<CacheETag>,
-    namespace: CacheNamespace,
-    diagnostics: Arc<CacheDiagnostics>,
-    #[cfg(test)]
-    barrier: Option<RestoreBarrier>,
-}
-
-impl PersistentRestore {
-    fn restore(&self) -> Option<CacheEntry> {
-        let started = Instant::now();
-        let prepared = {
-            let mut pack_file = self
-                .pack_file
-                .lock()
-                .expect("PackFile mutex should not be poisoned");
-            match self.namespace {
-                RESOLVE_CACHE_NAMESPACE => PreparedPersistentRecord::Resolve(
-                    pack_file.prepare_restore(&self.address, self.etag.as_ref())?,
-                ),
-                MODULE_BUILD_CACHE_NAMESPACE => PreparedPersistentRecord::ModuleBuild(
-                    pack_file.prepare_restore(&self.address, self.etag.as_ref())?,
-                ),
-                CODE_GENERATION_CACHE_NAMESPACE => PreparedPersistentRecord::CodeGeneration(
-                    pack_file.prepare_restore(&self.address, self.etag.as_ref())?,
-                ),
-                ASSET_RENDER_CACHE_NAMESPACE => PreparedPersistentRecord::AssetRender(
-                    pack_file.prepare_restore(&self.address, self.etag.as_ref())?,
-                ),
-                _ => return None,
-            }
-        };
-        #[cfg(test)]
-        if let Some(barrier) = &self.barrier {
-            barrier.entered.wait();
-            barrier.release.wait();
-        }
-
-        let restored = match prepared {
-            PreparedPersistentRecord::Resolve(prepared) => {
-                let record = ResolveRecord::try_from(prepared.decode()?).ok()?;
-                CacheEntry::new(CacheItemFamily::Resolve, self.cache_etag.clone(), record)
-            }
-            PreparedPersistentRecord::ModuleBuild(prepared) => {
-                let record = ModuleBuildRecord::try_from(prepared.decode()?).ok()?;
-                CacheEntry::new(
-                    CacheItemFamily::ModuleBuild,
-                    self.cache_etag.clone(),
-                    record,
-                )
-            }
-            PreparedPersistentRecord::CodeGeneration(prepared) => {
-                let record = prepared.decode()?.into_record_after_codec_validation();
-                CacheEntry::new(
-                    CacheItemFamily::CodeGeneration,
-                    self.cache_etag.clone(),
-                    record,
-                )
-            }
-            PreparedPersistentRecord::AssetRender(prepared) => {
-                let record = RenderedSource::try_from(prepared.decode()?).ok()?;
-                CacheEntry::new(
-                    CacheItemFamily::AssetRender,
-                    self.cache_etag.clone(),
-                    record,
-                )
-            }
-        };
-        self.diagnostics.profile(format!(
-            "restore items=1; deserialization items=1 duration_us={}",
-            started.elapsed().as_micros()
-        ));
-        Some(restored)
-    }
-
-    fn reads_from(&self, other: &Self) -> bool {
-        self.reader_generation == other.reader_generation
-            && Arc::ptr_eq(&self.pack_file, &other.pack_file)
-    }
-}
-
-#[derive(Debug)]
-struct PersistentTouch {
-    pack_file: Arc<Mutex<PackFile>>,
-    address: PackFileAddress,
-    etag: Option<PackFileETag>,
-    stamp: AccessStamp,
-}
-
-impl PersistentTouch {
-    fn apply(&self) -> bool {
-        self.pack_file
-            .lock()
-            .expect("PackFile mutex should not be poisoned")
-            .touch(&self.address, self.etag.as_ref(), self.stamp)
-    }
-}
-
-trait CacheLayer: fmt::Debug + Send + Sync {
-    fn lookup(&mut self, address: &CacheAddress, etag: Option<&CacheETag>) -> CacheLayerLookup;
-    fn store(&mut self, address: CacheAddress, entry: CacheEntry);
-    fn clear(&mut self) {}
-    fn prepare_persistent(
-        &mut self,
-        _guard: &PackFileGuardDto,
-        _build_inputs: &BTreeSet<PathBuf>,
-        _resolved_build_inputs: &BTreeSet<PathBuf>,
-        _automatic_build_inputs: &BTreeSet<PathBuf>,
-        _file_system_info: &FileSystemInfo,
-        _build_dependency_snapshot_strategy: SnapshotStrategy,
-        _resolve_build_dependency_snapshot_strategy: SnapshotStrategy,
-    ) -> bool {
-        false
-    }
-    fn plan_touch(
-        &self,
-        _address: &CacheAddress,
-        _etag: Option<&CacheETag>,
-        _stamp: AccessStamp,
-    ) -> Option<PersistentTouch> {
-        None
-    }
-    fn on_compilation_completed(&mut self) -> Vec<CacheItemFamily> {
-        Vec::new()
-    }
-    fn publish(
-        &mut self,
-        _guard: &PackFileGuardDto,
-        _stamp: AccessStamp,
-        _max_age: Duration,
-    ) -> io::Result<()> {
-        Ok(())
-    }
-    fn has_publication(&self) -> bool {
-        false
-    }
-    #[cfg(test)]
-    fn install_restore_barrier(&mut self, _barrier: RestoreBarrier) {}
-    #[allow(dead_code)]
-    fn evict(&mut self, address: &CacheAddress) -> bool;
-    #[cfg(test)]
-    fn entry_count(&self, family: CacheItemFamily) -> usize;
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum MemoryRetention {
-    Disabled,
-    Generations(u64),
-    Unbounded,
-}
-
-impl MemoryRetention {
-    fn from_options(options: &CacheOptions) -> Self {
-        match options.kind {
-            CacheKind::Disabled => Self::Disabled,
-            CacheKind::Memory | CacheKind::Filesystem => match options.max_memory_generations {
-                Some(0) => Self::Disabled,
-                Some(generations) => Self::Generations(generations),
-                None => Self::Unbounded,
-            },
-        }
-    }
-}
-
-#[derive(Debug, Clone)]
-struct MemoryCacheEntry {
-    entry: CacheEntry,
-    last_used_generation: u64,
-}
-
-#[derive(Debug)]
-struct MemoryCacheLayer {
-    entries: HashMap<CacheAddress, MemoryCacheEntry>,
-    retention: MemoryRetention,
-    completed_generation: u64,
-}
-
-impl MemoryCacheLayer {
-    fn new(retention: MemoryRetention) -> Self {
-        Self {
-            entries: HashMap::new(),
-            retention,
-            completed_generation: 0,
-        }
-    }
-
-    fn active_generation(&self) -> u64 {
-        self.completed_generation.saturating_add(1)
-    }
-}
-
-impl CacheLayer for MemoryCacheLayer {
-    fn lookup(&mut self, address: &CacheAddress, etag: Option<&CacheETag>) -> CacheLayerLookup {
-        if self.retention == MemoryRetention::Unbounded {
-            return self
-                .entries
-                .get(address)
-                .filter(|entry| entry.entry.etag.as_ref() == etag)
-                .map_or(CacheLayerLookup::Miss, |entry| {
-                    CacheLayerLookup::Hit(entry.entry.clone())
-                });
-        }
-        let active_generation = self.active_generation();
-        let Some(entry) = self
-            .entries
-            .get_mut(address)
-            .filter(|entry| entry.entry.etag.as_ref() == etag)
-        else {
-            return CacheLayerLookup::Miss;
-        };
-        entry.last_used_generation = active_generation;
-        CacheLayerLookup::Hit(entry.entry.clone())
-    }
-
-    fn store(&mut self, address: CacheAddress, entry: CacheEntry) {
-        self.entries.insert(
-            address,
-            MemoryCacheEntry {
-                entry,
-                last_used_generation: self.active_generation(),
-            },
-        );
-    }
-
-    fn on_compilation_completed(&mut self) -> Vec<CacheItemFamily> {
-        let MemoryRetention::Generations(limit) = self.retention else {
-            return Vec::new();
-        };
-        self.completed_generation = self.completed_generation.saturating_add(1);
-        let completed_generation = self.completed_generation;
-        let mut evicted = Vec::new();
-        self.entries.retain(|_, entry| {
-            let should_retain =
-                completed_generation.saturating_sub(entry.last_used_generation) < limit;
-            if !should_retain {
-                evicted.push(entry.entry.family);
-            }
-            should_retain
-        });
-        evicted
-    }
-
-    fn clear(&mut self) {
-        self.entries.clear();
-    }
-
-    fn evict(&mut self, address: &CacheAddress) -> bool {
-        self.entries.remove(address).is_some()
-    }
-
-    #[cfg(test)]
-    fn entry_count(&self, family: CacheItemFamily) -> usize {
-        self.entries
-            .values()
-            .filter(|entry| entry.entry.family == family)
-            .count()
-    }
-}
-
-#[derive(Debug)]
-struct PackFileCacheLayer {
-    root: Option<PathBuf>,
-    registry: CodecRegistry,
-    pack_file: Option<Arc<Mutex<PackFile>>>,
-    reader_generation: u64,
-    compression: PackFileCompression,
-    open_options: PackFileOpenOptions,
-    publication_base: PublicationBase,
-    active: bool,
-    pending: HashMap<CacheAddress, CacheEntry>,
-    diagnostics: Arc<CacheDiagnostics>,
-    _writer_marker: Option<CacheWriterMarker>,
-    #[cfg(test)]
-    restore_barrier: Option<RestoreBarrier>,
-}
-
-#[derive(Debug)]
-struct CacheWriterMarker {
-    path: PathBuf,
-    contents: String,
-}
-
-static NEXT_WRITER_TOKEN: AtomicU64 = AtomicU64::new(1);
-static PROCESS_START_ID: OnceLock<String> = OnceLock::new();
-
-impl CacheWriterMarker {
-    fn acquire(root: &Path, diagnostics: &CacheDiagnostics) -> Option<Self> {
-        if let Err(error) = fs::create_dir_all(root) {
-            diagnostics.warn(format!(
-                "single-writer diagnostic unavailable at {}: {error}; continuing best-effort",
-                root.display()
-            ));
-            return None;
-        }
-        let path = root.join(".unpack-writer");
-        let pid = std::process::id();
-        let start = current_process_start_id();
-        let token = NEXT_WRITER_TOKEN.fetch_add(1, Ordering::Relaxed);
-        let contents = format!("UNPACK-WRITER-1\npid={pid}\nstart={start}\ntoken={token}\n");
-        for _ in 0..2 {
-            match fs::OpenOptions::new()
-                .write(true)
-                .create_new(true)
-                .open(&path)
-            {
-                Ok(mut file) => {
-                    use std::io::Write;
-                    if file.write_all(contents.as_bytes()).is_ok() {
-                        return Some(Self { path, contents });
-                    }
-                    let _ = fs::remove_file(&path);
-                    return None;
-                }
-                Err(error) if error.kind() == io::ErrorKind::AlreadyExists => {
-                    let existing = fs::read_to_string(&path).unwrap_or_default();
-                    if writer_marker_is_live(&existing) {
-                        diagnostics.warn(format!(
-                            "detected another live writer for Persistent Cache location {}; continuing best-effort without a cross-process lock or merge protocol; contract=trusted-local,linux-supported,single-writer",
-                            root.display()
-                        ));
-                        return None;
-                    }
-                    let _ = fs::remove_file(&path);
-                }
-                Err(error) => {
-                    diagnostics.warn(format!(
-                        "single-writer diagnostic unavailable at {}: {error}; continuing best-effort",
-                        root.display()
-                    ));
-                    return None;
-                }
-            }
-        }
-        None
-    }
-}
-
-impl Drop for CacheWriterMarker {
-    fn drop(&mut self) {
-        if fs::read_to_string(&self.path).ok().as_deref() == Some(self.contents.as_str()) {
-            let _ = fs::remove_file(&self.path);
-        }
-    }
-}
-
-fn current_process_start_id() -> &'static str {
-    PROCESS_START_ID.get_or_init(|| {
-        linux_process_start_id(std::process::id()).unwrap_or_else(|| {
-            SystemTime::now()
-                .duration_since(UNIX_EPOCH)
-                .unwrap_or_default()
-                .as_nanos()
-                .to_string()
-        })
-    })
-}
-
-fn linux_process_start_id(pid: u32) -> Option<String> {
-    let stat = fs::read_to_string(format!("/proc/{pid}/stat")).ok()?;
-    let fields = stat
-        .rsplit_once(") ")?
-        .1
-        .split_whitespace()
-        .collect::<Vec<_>>();
-    fields.get(19).map(|value| (*value).to_string())
-}
-
-fn writer_marker_is_live(contents: &str) -> bool {
-    let value = |prefix: &str| contents.lines().find_map(|line| line.strip_prefix(prefix));
-    let Some(pid) = value("pid=").and_then(|pid| pid.parse::<u32>().ok()) else {
-        return false;
-    };
-    let Some(start) = value("start=") else {
-        return false;
-    };
-    if pid == std::process::id() {
-        return start == current_process_start_id();
-    }
-    linux_process_start_id(pid).is_some_and(|candidate| candidate == start)
-        || !cfg!(target_os = "linux")
-}
-
-impl PackFileCacheLayer {
-    fn open(options: &CacheOptions, diagnostics: Arc<CacheDiagnostics>) -> Self {
-        let registry = persistent_codec_registry();
-        let root = options.cache_location.clone();
-        let compression = options.compression.into();
-        let open_options = PackFileOpenOptions::new(options.allow_collecting_memory);
-        let pack_file = root.as_ref().map(|root| {
-            Arc::new(Mutex::new(PackFile::open_with_options(
-                root,
-                registry.clone(),
-                open_options,
-            )))
-        });
-        let has_standalone_guard = pack_file.as_ref().is_some_and(|pack_file| {
-            pack_file
-                .lock()
-                .expect("PackFile mutex should not be poisoned")
-                .guard()
-                .is_some_and(|guard| {
-                    guard.build_dependencies.entries.is_empty()
-                        && guard.resolve_build_dependencies.entries.is_empty()
-                })
-        });
-        let publication_base = if has_standalone_guard {
-            PublicationBase::PreserveEntries {
-                expected_revision: pack_file
-                    .as_ref()
-                    .expect("standalone PackFile should be open")
-                    .lock()
-                    .expect("PackFile mutex should not be poisoned")
-                    .revision(),
-            }
-        } else {
-            PublicationBase::ReplaceAll
-        };
-        let writer_marker = if options.readonly {
-            None
-        } else {
-            root.as_deref()
-                .and_then(|root| CacheWriterMarker::acquire(root, &diagnostics))
-        };
-        Self {
-            root,
-            registry,
-            pack_file,
-            reader_generation: 0,
-            compression,
-            open_options,
-            publication_base,
-            // Compiler preparation validates non-empty Build Dependency guards before
-            // work starts.  Standalone cache users have no such guard, and may safely
-            // restore the legacy empty-guard PackFile directly.
-            active: true,
-            pending: HashMap::new(),
-            diagnostics,
-            _writer_marker: writer_marker,
-            #[cfg(test)]
-            restore_barrier: None,
-        }
-    }
-
-    fn publish(
-        &mut self,
-        guard: PackFileGuardDto,
-        stamp: AccessStamp,
-        max_age: Duration,
-    ) -> io::Result<()> {
-        let Some(root) = self.root.as_ref() else {
-            self.pending.clear();
-            return Ok(());
-        };
-        let before_entries = self.pack_file.as_ref().map_or(0, |pack_file| {
-            pack_file
-                .lock()
-                .expect("PackFile mutex should not be poisoned")
-                .entry_count()
-        });
-        let queued_items = self.pending.len();
-        let serialization_started = Instant::now();
-        let mut batch = PackFileWriteBatch::new();
-        if let Some(pack_file) = &self.pack_file {
-            pack_file
-                .lock()
-                .expect("PackFile mutex should not be poisoned")
-                .copy_access_updates_to(&mut batch);
-        }
-        for (address, entry) in &self.pending {
-            let pack_address = address.to_pack_file_address();
-            let pack_etag = entry.etag.as_ref().map(CacheETag::to_pack_file_etag);
-            match entry.family {
-                CacheItemFamily::Resolve => {
-                    let record = entry.value::<ResolveRecord>().ok_or_else(|| {
-                        io::Error::new(
-                            io::ErrorKind::InvalidData,
-                            "Resolve Cache Item contains an unexpected value",
-                        )
-                    })?;
-                    let dto = ResolveRecordDto::try_from(record.as_ref())?;
-                    batch.insert(&self.registry, pack_address, pack_etag, dto)?;
-                }
-                CacheItemFamily::ModuleBuild => {
-                    let record = entry.value::<ModuleBuildRecord>().ok_or_else(|| {
-                        io::Error::new(
-                            io::ErrorKind::InvalidData,
-                            "Module Build Cache Item contains an unexpected value",
-                        )
-                    })?;
-                    let dto = ModuleBuildRecordDto::try_from(record.as_ref())?;
-                    batch.insert(&self.registry, pack_address, pack_etag, dto)?;
-                }
-                CacheItemFamily::CodeGeneration => {
-                    let record = entry.value::<CodeGenerationRecord>().ok_or_else(|| {
-                        io::Error::new(
-                            io::ErrorKind::InvalidData,
-                            "Code Generation Cache Item contains an unexpected value",
-                        )
-                    })?;
-                    let dto = CodeGenerationRecordDto::from(record.as_ref());
-                    batch.insert(&self.registry, pack_address, pack_etag, dto)?;
-                }
-                CacheItemFamily::AssetRender => {
-                    let record = entry.value::<RenderedSource>().ok_or_else(|| {
-                        io::Error::new(
-                            io::ErrorKind::InvalidData,
-                            "Asset Render Cache Item contains an unexpected value",
-                        )
-                    })?;
-                    let dto = AssetRenderRecordDto::from(record.as_ref());
-                    batch.insert(&self.registry, pack_address, pack_etag, dto)?;
-                }
-            }
-        }
-        self.diagnostics.profile(format!(
-            "serialization items={queued_items} duration_us={}",
-            serialization_started.elapsed().as_micros()
-        ));
-        let store_started = Instant::now();
-        PackFile::publish_batch_with_options(
-            root,
-            Some(guard),
-            self.publication_base,
-            batch,
-            PackFilePublicationOptions::new(
-                PackFileRetention::new(stamp, max_age),
-                self.compression,
-            ),
-        )?;
-        self.pending.clear();
-        let pack_file = PackFile::open_with_options(root, self.registry.clone(), self.open_options);
-        let after_entries = pack_file.entry_count();
-        self.diagnostics.profile(format!(
-            "store items={queued_items} duration_us={}; garbage collection removed_items={}; merge packs=0; split packs=0; compaction packs=0",
-            store_started.elapsed().as_micros(),
-            before_entries.saturating_sub(after_entries)
-        ));
-        self.publication_base = PublicationBase::PreserveEntries {
-            expected_revision: pack_file.revision(),
-        };
-        self.pack_file = Some(Arc::new(Mutex::new(pack_file)));
-        self.reader_generation = self.reader_generation.wrapping_add(1);
-        self.active = true;
-        Ok(())
-    }
-}
-
-impl CacheLayer for PackFileCacheLayer {
-    fn lookup(&mut self, address: &CacheAddress, etag: Option<&CacheETag>) -> CacheLayerLookup {
-        if let Some(entry) = self
-            .pending
-            .get(address)
-            .filter(|entry| entry.etag.as_ref() == etag)
-        {
-            return CacheLayerLookup::Hit(entry.clone());
-        }
-        if !self.active {
-            return CacheLayerLookup::Miss;
-        }
-        let Some(pack_file) = self.pack_file.as_ref() else {
-            return CacheLayerLookup::Miss;
-        };
-        CacheLayerLookup::Deferred(PersistentRestore {
-            pack_file: Arc::clone(pack_file),
-            reader_generation: self.reader_generation,
-            address: address.to_pack_file_address(),
-            etag: etag.map(CacheETag::to_pack_file_etag),
-            cache_etag: etag.cloned(),
-            namespace: address.namespace,
-            diagnostics: Arc::clone(&self.diagnostics),
-            #[cfg(test)]
-            barrier: self.restore_barrier.take(),
-        })
-    }
-
-    fn store(&mut self, address: CacheAddress, entry: CacheEntry) {
-        self.diagnostics.profile("store queued_items=1");
-        self.pending.insert(address, entry);
-    }
-
-    fn clear(&mut self) {
-        self.pending.clear();
-        self.active = false;
-        self.reader_generation = self.reader_generation.wrapping_add(1);
-        self.publication_base = PublicationBase::ReplaceAll;
-    }
-
-    fn prepare_persistent(
-        &mut self,
-        guard: &PackFileGuardDto,
-        build_inputs: &BTreeSet<PathBuf>,
-        resolved_build_inputs: &BTreeSet<PathBuf>,
-        automatic_build_inputs: &BTreeSet<PathBuf>,
-        file_system_info: &FileSystemInfo,
-        build_dependency_snapshot_strategy: SnapshotStrategy,
-        resolve_build_dependency_snapshot_strategy: SnapshotStrategy,
-    ) -> bool {
-        let build_strategy = if !build_inputs.is_empty()
-            && build_inputs
-                .iter()
-                .all(|path| automatic_build_inputs.contains(path))
-        {
-            SnapshotStrategy::timestamp()
-        } else {
-            build_dependency_snapshot_strategy
-        };
-        let active = self.pack_file.as_ref().is_some_and(|pack_file| {
-            pack_file
-                .lock()
-                .expect("PackFile mutex should not be poisoned")
-                .guard()
-                .is_some_and(|candidate| {
-                    let build_dependencies =
-                        Snapshot::try_from(candidate.build_dependencies.clone()).ok();
-                    let resolve_build_dependencies =
-                        Snapshot::try_from(candidate.resolve_build_dependencies.clone()).ok();
-                    candidate.version == guard.version
-                        && build_dependencies.is_some_and(|snapshot| {
-                            snapshot.has_exact_paths(build_inputs.iter().cloned())
-                                && file_system_info
-                                    .is_snapshot_valid_sync(&snapshot, build_strategy)
-                        })
-                        && resolve_build_dependencies.is_some_and(|snapshot| {
-                            file_system_info.is_snapshot_valid_sync(
-                                &snapshot,
-                                resolve_build_dependency_snapshot_strategy,
-                            ) || snapshot.has_valid_paths_sync(
-                                resolved_build_inputs.iter().cloned(),
-                                resolve_build_dependency_snapshot_strategy,
-                                file_system_info,
-                            )
-                        })
-                })
-        });
-        self.reader_generation = self.reader_generation.wrapping_add(1);
-        self.active = active;
-        self.publication_base = if active {
-            PublicationBase::PreserveEntries {
-                expected_revision: self
-                    .pack_file
-                    .as_ref()
-                    .expect("active PackFile should be open")
-                    .lock()
-                    .expect("PackFile mutex should not be poisoned")
-                    .revision(),
-            }
-        } else {
-            PublicationBase::ReplaceAll
-        };
-        self.pack_file
-            .as_ref()
-            .and_then(|pack_file| {
-                pack_file
-                    .lock()
-                    .expect("PackFile mutex should not be poisoned")
-                    .guard()
-                    .cloned()
-            })
-            .as_ref()
-            != Some(guard)
-    }
-
-    fn plan_touch(
-        &self,
-        address: &CacheAddress,
-        etag: Option<&CacheETag>,
-        stamp: AccessStamp,
-    ) -> Option<PersistentTouch> {
-        if !self.active {
-            return None;
-        }
-        self.pack_file.as_ref().map(|pack_file| PersistentTouch {
-            pack_file: Arc::clone(pack_file),
-            address: address.to_pack_file_address(),
-            etag: etag.map(CacheETag::to_pack_file_etag),
-            stamp,
-        })
-    }
-
-    fn publish(
-        &mut self,
-        guard: &PackFileGuardDto,
-        stamp: AccessStamp,
-        max_age: Duration,
-    ) -> io::Result<()> {
-        PackFileCacheLayer::publish(self, guard.clone(), stamp, max_age)
-    }
-
-    fn has_publication(&self) -> bool {
-        self.active
-    }
-
-    #[cfg(test)]
-    fn install_restore_barrier(&mut self, barrier: RestoreBarrier) {
-        self.restore_barrier = Some(barrier);
-    }
-
-    fn evict(&mut self, address: &CacheAddress) -> bool {
-        self.pending.remove(address).is_some()
-    }
-
-    #[cfg(test)]
-    fn entry_count(&self, family: CacheItemFamily) -> usize {
-        self.pending
-            .values()
-            .filter(|entry| entry.family == family)
-            .count()
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum CacheLayerKind {
-    Memory,
-    Persistent,
-}
-
-#[derive(Debug)]
-struct CacheLayerSlot {
-    kind: CacheLayerKind,
-    writable: bool,
-    layer: Box<dyn CacheLayer>,
-}
-
-#[derive(Debug)]
-struct Cache {
-    layers: Vec<CacheLayerSlot>,
-    work: CacheWorkCounters,
-}
-
-struct CacheRestore {
-    layer_index: usize,
-    restore: PersistentRestore,
-}
-
-enum CacheGet<V> {
-    Ready {
-        value: Option<Arc<V>>,
-        persistent_access_changed: bool,
-    },
-    Deferred(CacheRestore),
-}
-
-impl<V> CacheGet<V> {
-    fn persistent_access_changed(&self) -> bool {
-        matches!(
-            self,
-            Self::Ready {
-                persistent_access_changed: true,
-                ..
-            }
-        )
-    }
-}
-
-impl Cache {
-    fn from_options(options: &CacheOptions, diagnostics: Arc<CacheDiagnostics>) -> Self {
-        let mut layers = Vec::new();
-        let memory_retention = MemoryRetention::from_options(options);
-        if memory_retention != MemoryRetention::Disabled {
-            layers.push(CacheLayerSlot {
-                kind: CacheLayerKind::Memory,
-                writable: true,
-                layer: Box::new(MemoryCacheLayer::new(memory_retention)),
-            });
-        }
-        if options.kind == CacheKind::Filesystem {
-            layers.push(CacheLayerSlot {
-                kind: CacheLayerKind::Persistent,
-                writable: !options.readonly,
-                layer: Box::new(PackFileCacheLayer::open(options, diagnostics)),
-            });
-        }
-        Self {
-            layers,
-            work: CacheWorkCounters::default(),
-        }
-    }
-
-    fn begin_get<V>(
-        &mut self,
-        family: CacheItemFamily,
-        address: &CacheAddress,
-        etag: Option<&CacheETag>,
-        stamp: AccessStamp,
-    ) -> CacheGet<V>
-    where
-        V: Send + Sync + 'static,
-    {
-        for index in 0..self.layers.len() {
-            match self.layers[index].layer.lookup(address, etag) {
-                CacheLayerLookup::Hit(entry) => {
-                    if let Some((value, persistent_access_changed)) =
-                        self.use_entry(family, address, etag, stamp, index, entry)
-                    {
-                        return CacheGet::Ready {
-                            value: Some(value),
-                            persistent_access_changed,
-                        };
-                    }
-                }
-                CacheLayerLookup::Deferred(restore) => {
-                    return CacheGet::Deferred(CacheRestore {
-                        layer_index: index,
-                        restore,
-                    });
-                }
-                CacheLayerLookup::Miss => {}
-            }
-        }
-
-        self.work.for_family_mut(family).misses += 1;
-        CacheGet::Ready {
-            value: None,
-            persistent_access_changed: false,
-        }
-    }
-
-    fn finish_restore<V>(
-        &mut self,
-        family: CacheItemFamily,
-        address: &CacheAddress,
-        etag: Option<&CacheETag>,
-        stamp: AccessStamp,
-        plan: CacheRestore,
-        restored: Option<CacheEntry>,
-    ) -> CacheGet<V>
-    where
-        V: Send + Sync + 'static,
-    {
-        for index in 0..plan.layer_index {
-            if let CacheLayerLookup::Hit(entry) = self.layers[index].layer.lookup(address, etag)
-                && let Some((value, persistent_access_changed)) =
-                    self.use_entry(family, address, etag, stamp, index, entry)
-            {
-                return CacheGet::Ready {
-                    value: Some(value),
-                    persistent_access_changed,
-                };
-            }
-        }
-
-        match self.layers[plan.layer_index].layer.lookup(address, etag) {
-            CacheLayerLookup::Hit(entry) => {
-                if let Some((value, persistent_access_changed)) =
-                    self.use_entry(family, address, etag, stamp, plan.layer_index, entry)
-                {
-                    return CacheGet::Ready {
-                        value: Some(value),
-                        persistent_access_changed,
-                    };
-                }
-            }
-            CacheLayerLookup::Deferred(current) if current.reads_from(&plan.restore) => {
-                if let Some(entry) = restored
-                    && let Some((value, persistent_access_changed)) =
-                        self.use_entry(family, address, etag, stamp, plan.layer_index, entry)
-                {
-                    return CacheGet::Ready {
-                        value: Some(value),
-                        persistent_access_changed,
-                    };
-                }
-            }
-            CacheLayerLookup::Deferred(current) => {
-                return CacheGet::Deferred(CacheRestore {
-                    layer_index: plan.layer_index,
-                    restore: current,
-                });
-            }
-            CacheLayerLookup::Miss => {}
-        }
-
-        self.work.for_family_mut(family).misses += 1;
-        CacheGet::Ready {
-            value: None,
-            persistent_access_changed: false,
-        }
-    }
-
-    fn use_entry<V>(
-        &mut self,
-        family: CacheItemFamily,
-        address: &CacheAddress,
-        etag: Option<&CacheETag>,
-        stamp: AccessStamp,
-        layer_index: usize,
-        entry: CacheEntry,
-    ) -> Option<(Arc<V>, bool)>
-    where
-        V: Send + Sync + 'static,
-    {
-        let value = entry.value::<V>()?;
-        if layer_index > 0 {
-            for earlier in &mut self.layers[..layer_index] {
-                if earlier.writable {
-                    earlier.layer.store(address.clone(), entry.clone());
-                }
-            }
-            self.work.for_family_mut(family).restores += 1;
-        }
-        self.work.for_family_mut(family).hits += 1;
-        let persistent_access_changed = self
-            .layers
-            .iter()
-            .filter(|slot| slot.kind == CacheLayerKind::Persistent && slot.writable)
-            .find_map(|slot| slot.layer.plan_touch(address, etag, stamp))
-            .is_some_and(|touch| touch.apply());
-        Some((value, persistent_access_changed))
-    }
-
-    fn store<V>(
-        &mut self,
-        family: CacheItemFamily,
-        address: CacheAddress,
-        etag: Option<CacheETag>,
-        value: V,
-    ) -> bool
-    where
-        V: Send + Sync + 'static,
-    {
-        let entry = CacheEntry::new(family, etag, value);
-        let mut stored_persistently = false;
-        for slot in &mut self.layers {
-            if !slot.writable {
-                continue;
-            }
-            slot.layer.store(address.clone(), entry.clone());
-            stored_persistently |= slot.kind == CacheLayerKind::Persistent;
-        }
-        self.work.for_family_mut(family).stores += 1;
-        stored_persistently
-    }
-
-    #[cfg(test)]
-    fn evict_memory(&mut self, family: CacheItemFamily, address: &CacheAddress) {
-        let evicted = self
-            .layers
-            .iter_mut()
-            .filter(|slot| slot.kind == CacheLayerKind::Memory)
-            .any(|slot| slot.layer.evict(address));
-        if evicted {
-            self.work.for_family_mut(family).evictions += 1;
-        }
-    }
-
-    #[cfg(test)]
-    fn entry_count(&self, family: CacheItemFamily) -> usize {
-        self.layers
-            .iter()
-            .find(|slot| slot.kind == CacheLayerKind::Memory)
-            .map(|slot| slot.layer.entry_count(family))
-            .unwrap_or_default()
-    }
-
-    fn work_counters(&self) -> CacheWorkCounters {
-        self.work
-    }
-
-    fn clear(&mut self) {
-        for slot in &mut self.layers {
-            slot.layer.clear();
-        }
-    }
-
-    fn prepare_persistent(
-        &mut self,
-        guard: &PackFileGuardDto,
-        build_inputs: &BTreeSet<PathBuf>,
-        resolved_build_inputs: &BTreeSet<PathBuf>,
-        automatic_build_inputs: &BTreeSet<PathBuf>,
-        file_system_info: &FileSystemInfo,
-        build_dependency_snapshot_strategy: SnapshotStrategy,
-        resolve_build_dependency_snapshot_strategy: SnapshotStrategy,
-    ) -> bool {
-        self.layers
-            .iter_mut()
-            .find(|slot| slot.kind == CacheLayerKind::Persistent)
-            .is_some_and(|slot| {
-                slot.layer.prepare_persistent(
-                    guard,
-                    build_inputs,
-                    resolved_build_inputs,
-                    automatic_build_inputs,
-                    file_system_info,
-                    build_dependency_snapshot_strategy,
-                    resolve_build_dependency_snapshot_strategy,
-                )
-            })
-    }
-
-    fn on_compilation_completed(&mut self) {
-        for slot in &mut self.layers {
-            for family in slot.layer.on_compilation_completed() {
-                self.work.for_family_mut(family).evictions += 1;
-            }
-        }
-    }
-
-    fn publish_persistent(
-        &mut self,
-        guard: PackFileGuardDto,
-        stamp: AccessStamp,
-        max_age: Duration,
-    ) -> io::Result<()> {
-        let Some(slot) = self
-            .layers
-            .iter_mut()
-            .find(|slot| slot.kind == CacheLayerKind::Persistent)
-        else {
-            return Ok(());
-        };
-        slot.layer.publish(&guard, stamp, max_age)
-    }
-
-    fn has_persistent_publication(&self) -> bool {
-        self.layers
-            .iter()
-            .find(|slot| slot.kind == CacheLayerKind::Persistent)
-            .is_some_and(|slot| slot.layer.has_publication())
-    }
-
-    #[cfg(test)]
-    fn install_restore_barrier(&mut self, barrier: RestoreBarrier) {
-        if let Some(slot) = self
-            .layers
-            .iter_mut()
-            .find(|slot| slot.kind == CacheLayerKind::Persistent)
-        {
-            slot.layer.install_restore_barrier(barrier);
-        }
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Hash)]
-pub(crate) struct ResolveRequest {
-    context: PathBuf,
-    request: String,
-}
-
-impl ResolveRequest {
-    pub(crate) fn new(context: impl Into<PathBuf>, request: impl Into<String>) -> Self {
-        Self {
-            context: context.into(),
-            request: request.into(),
-        }
-    }
-}
-
-impl CacheKey for ResolveRequest {
-    fn cache_identifier(&self) -> CacheIdentifier {
-        CacheIdentifier::from_parts([
-            self.context.as_os_str().as_encoded_bytes().to_vec(),
-            self.request.as_bytes().to_vec(),
-        ])
-    }
-}
-
-impl CacheKey for ModuleIdentity {
-    fn cache_identifier(&self) -> CacheIdentifier {
-        let mut parts = vec![
-            match self.module_type {
-                crate::ModuleType::JavaScriptAuto => b"javascript/auto".to_vec(),
-            },
-            self.resource.as_os_str().as_encoded_bytes().to_vec(),
-            optional_identifier_part(self.query.as_deref()),
-            optional_identifier_part(self.fragment.as_deref()),
-            optional_identifier_part(self.layer.as_deref()),
-            (self.loaders.len() as u64).to_le_bytes().to_vec(),
-        ];
-        parts.extend(self.loaders.iter().map(|loader| loader.as_bytes().to_vec()));
-        CacheIdentifier::from_parts(parts)
-    }
-}
-
-fn optional_identifier_part(value: Option<&str>) -> Vec<u8> {
-    match value {
-        Some(value) => {
-            let mut bytes = Vec::with_capacity(value.len() + 1);
-            bytes.push(1);
-            bytes.extend_from_slice(value.as_bytes());
-            bytes
-        }
-        None => vec![0],
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub(crate) struct ResolveRecord {
-    identity: ModuleIdentity,
-    resource: PathBuf,
-    file_dependencies: BTreeSet<PathBuf>,
-    context_dependencies: BTreeSet<PathBuf>,
-    missing_dependencies: BTreeSet<PathBuf>,
-    snapshot: Snapshot,
-}
-
-impl ResolveRecord {
-    #[cfg(test)]
-    pub(crate) async fn new(
-        identity: ModuleIdentity,
-        resource: PathBuf,
-        file_dependencies: BTreeSet<PathBuf>,
-        context_dependencies: BTreeSet<PathBuf>,
-        missing_dependencies: BTreeSet<PathBuf>,
-        file_system_info: &FileSystemInfo,
-        strategy: SnapshotStrategy,
-    ) -> crate::Result<Self> {
-        let snapshot = file_system_info
-            .create_resolve_snapshot(
-                file_dependencies.iter().cloned(),
-                context_dependencies.iter().cloned(),
-                missing_dependencies.iter().cloned(),
-                strategy,
-            )
-            .await?;
-        Ok(Self {
-            identity,
-            resource,
-            file_dependencies,
-            context_dependencies,
-            missing_dependencies,
-            snapshot,
-        })
-    }
-
-    pub(crate) async fn new_with_cache(
-        identity: ModuleIdentity,
-        resource: PathBuf,
-        file_dependencies: BTreeSet<PathBuf>,
-        context_dependencies: BTreeSet<PathBuf>,
-        missing_dependencies: BTreeSet<PathBuf>,
-        file_system_info: &FileSystemInfo,
-        strategy: SnapshotStrategy,
-        snapshot_cache: &SnapshotCache,
-    ) -> crate::Result<Self> {
-        let snapshot = file_system_info
-            .create_resolve_snapshot_with_cache(
-                file_dependencies.iter().cloned(),
-                context_dependencies.iter().cloned(),
-                missing_dependencies.iter().cloned(),
-                strategy,
-                snapshot_cache,
-            )
-            .await?;
-        Ok(Self {
-            identity,
-            resource,
-            file_dependencies,
-            context_dependencies,
-            missing_dependencies,
-            snapshot,
-        })
-    }
-
-    pub(crate) fn identity(&self) -> &ModuleIdentity {
-        &self.identity
-    }
-
-    pub(crate) fn resource(&self) -> &Path {
-        &self.resource
-    }
-
-    pub(crate) fn file_dependencies(&self) -> &BTreeSet<PathBuf> {
-        &self.file_dependencies
-    }
-
-    pub(crate) fn context_dependencies(&self) -> &BTreeSet<PathBuf> {
-        &self.context_dependencies
-    }
-
-    pub(crate) fn missing_dependencies(&self) -> &BTreeSet<PathBuf> {
-        &self.missing_dependencies
-    }
-
-    pub(crate) fn snapshot(&self) -> &Snapshot {
-        &self.snapshot
-    }
-
-    pub(crate) fn from_persistent_parts(
-        identity: ModuleIdentity,
-        resource: PathBuf,
-        file_dependencies: BTreeSet<PathBuf>,
-        context_dependencies: BTreeSet<PathBuf>,
-        missing_dependencies: BTreeSet<PathBuf>,
-        snapshot: Snapshot,
-    ) -> Self {
-        Self {
-            identity,
-            resource,
-            file_dependencies,
-            context_dependencies,
-            missing_dependencies,
-            snapshot,
-        }
-    }
-
-    #[cfg(test)]
-    pub(crate) async fn is_valid(
-        &self,
-        file_system_info: &FileSystemInfo,
-        strategy: SnapshotStrategy,
-    ) -> bool {
-        file_system_info
-            .is_snapshot_valid(&self.snapshot, strategy)
-            .await
-    }
-
-    pub(crate) async fn is_valid_with_cache(
-        &self,
-        file_system_info: &FileSystemInfo,
-        strategy: SnapshotStrategy,
-        snapshot_cache: &SnapshotCache,
-    ) -> bool {
-        file_system_info
-            .is_snapshot_valid_with_cache(&self.snapshot, strategy, snapshot_cache)
-            .await
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub(crate) struct ModuleBuildRecord {
-    built_content: Arc<BuiltModuleContent>,
-    snapshot: Snapshot,
-}
-
-impl ModuleBuildRecord {
-    pub(crate) fn new(built_content: Arc<BuiltModuleContent>, snapshot: Snapshot) -> Self {
-        Self {
-            built_content,
-            snapshot,
-        }
-    }
-
-    pub(crate) fn parsed(&self) -> &ParsedModule {
-        self.built_content.parsed()
-    }
-
-    pub(crate) fn built_content(&self) -> &Arc<BuiltModuleContent> {
-        &self.built_content
-    }
-
-    pub(crate) fn persistent_parts(&self) -> (&ParsedModule, &str, Option<u64>) {
-        (
-            self.built_content.parsed(),
-            self.built_content.source(),
-            Some(self.built_content.source_hash()),
-        )
-    }
-
-    pub(crate) fn from_persistent_parts(
-        parsed: ParsedModule,
-        source: String,
-        source_hash: u64,
-        snapshot: Snapshot,
-    ) -> Self {
-        Self {
-            built_content: Arc::new(BuiltModuleContent::from_persistent_parts(
-                parsed,
-                source,
-                source_hash,
-            )),
-            snapshot,
-        }
-    }
-
-    pub(crate) fn snapshot(&self) -> &Snapshot {
-        &self.snapshot
-    }
-
-    pub(crate) async fn is_valid_with_cache(
-        &self,
-        file_system_info: &FileSystemInfo,
-        strategy: SnapshotStrategy,
-        snapshot_cache: &SnapshotCache,
-    ) -> bool {
-        file_system_info
-            .is_snapshot_valid_with_cache(&self.snapshot, strategy, snapshot_cache)
-            .await
-    }
-}
-
 impl BuildCache {
     pub(crate) fn new(options: CacheOptions, snapshot_options: SnapshotOptions) -> Self {
         Self::new_with_clock_inner(options, snapshot_options, Arc::new(SystemCacheClock))
@@ -1968,15 +397,16 @@ impl BuildCache {
         if publication.persistent_guard.is_some() && !previous_build_inputs_are_valid {
             cache.clear();
         }
-        let guard_changed = cache.prepare_persistent(
-            &guard,
-            &build_inputs,
-            &resolved_build_inputs,
-            &automatic_build_inputs,
-            &self.build_dependency_file_system_info,
-            self.build_dependency_snapshot_strategy,
-            self.resolve_build_dependency_snapshot_strategy,
-        );
+        let guard_changed = cache.prepare_persistent(PersistentCachePreparation {
+            guard: &guard,
+            build_inputs: &build_inputs,
+            resolved_build_inputs: &resolved_build_inputs,
+            automatic_build_inputs: &automatic_build_inputs,
+            file_system_info: &self.build_dependency_file_system_info,
+            build_dependency_snapshot_strategy: self.build_dependency_snapshot_strategy,
+            resolve_build_dependency_snapshot_strategy: self
+                .resolve_build_dependency_snapshot_strategy,
+        });
         publication.persistent_guard = Some(guard);
         if guard_changed {
             self.inner.mark_dirty();
@@ -2096,7 +526,7 @@ impl BuildCache {
             .lock()
             .expect("build cache data mutex should not be poisoned");
         let work = cache.work_counters();
-        let resolve = work.for_family(CacheItemFamily::Resolve);
+        let resolve: CacheItemWork = work.for_family(CacheItemFamily::Resolve);
         let module = work.for_family(CacheItemFamily::ModuleBuild);
         BuildCacheStats {
             resolve_entries: cache.entry_count(CacheItemFamily::Resolve),
@@ -2118,7 +548,7 @@ impl BuildCache {
 
     pub(crate) fn trace_work_counters(&self) {
         let work = self.work_counters();
-        let resolve = work.for_family(CacheItemFamily::Resolve);
+        let resolve: CacheItemWork = work.for_family(CacheItemFamily::Resolve);
         let module = work.for_family(CacheItemFamily::ModuleBuild);
         let code_generation = work.for_family(CacheItemFamily::CodeGeneration);
         let asset_render = work.for_family(CacheItemFamily::AssetRender);
@@ -2155,131 +585,10 @@ impl BuildCache {
             .expect("build cache data mutex should not be poisoned")
             .on_compilation_completed();
     }
-}
 
-impl<K, V> CacheFacade<K, V>
-where
-    K: CacheKey,
-    V: Send + Sync + 'static,
-{
-    pub(crate) fn is_enabled(&self) -> bool {
-        self.build_cache.options.kind != CacheKind::Disabled
-    }
-
-    #[allow(dead_code)]
-    pub(crate) fn namespace(&self) -> CacheNamespace {
-        self.namespace
-    }
-
-    pub(crate) fn get(&self, key: &K, etag: Option<&CacheETag>) -> Option<Arc<V>> {
-        if !self.is_enabled() {
-            return None;
-        }
-
-        let address = CacheAddress {
-            namespace: self.namespace,
-            identifier: key.cache_identifier(),
-        };
-        // Access timestamps only matter for the filesystem layer. Avoid a clock
-        // syscall for memory caches and read-only persistent caches, which never
-        // record access updates.
-        let stamp = if self.build_cache.is_writable_filesystem_cache() {
-            self.build_cache.clock.now()
-        } else {
-            AccessStamp::from_millis(0)
-        };
-
-        let mut result = {
-            let mut cache = self
-                .build_cache
-                .inner
-                .cache
-                .lock()
-                .expect("build cache data mutex should not be poisoned");
-            let result = cache.begin_get(self.family, &address, etag, stamp);
-            if result.persistent_access_changed() {
-                self.build_cache.inner.mark_dirty();
-            }
-            result
-        };
-
-        loop {
-            match result {
-                CacheGet::Ready { value, .. } => return value,
-                CacheGet::Deferred(plan) => {
-                    let restored = plan.restore.restore();
-                    result = {
-                        let mut cache = self
-                            .build_cache
-                            .inner
-                            .cache
-                            .lock()
-                            .expect("build cache data mutex should not be poisoned");
-                        let result = cache.finish_restore(
-                            self.family,
-                            &address,
-                            etag,
-                            stamp,
-                            plan,
-                            restored,
-                        );
-                        if result.persistent_access_changed() {
-                            self.build_cache.inner.mark_dirty();
-                        }
-                        result
-                    };
-                }
-            }
-        }
-    }
-
-    pub(crate) fn store(&self, key: K, etag: Option<CacheETag>, value: V) {
-        if !self.is_enabled() {
-            return;
-        }
-
-        let address = CacheAddress {
-            namespace: self.namespace,
-            identifier: key.cache_identifier(),
-        };
-        let mut cache = self
-            .build_cache
-            .inner
-            .cache
-            .lock()
-            .expect("build cache data mutex should not be poisoned");
-        if cache.store(self.family, address, etag, value) {
-            self.build_cache.inner.mark_dirty();
-        }
-    }
-
-    #[cfg(test)]
-    pub(crate) fn evict_memory(&self, key: &K) {
-        let address = CacheAddress {
-            namespace: self.namespace,
-            identifier: key.cache_identifier(),
-        };
-        self.build_cache
-            .inner
-            .cache
-            .lock()
-            .expect("build cache data mutex should not be poisoned")
-            .evict_memory(self.family, &address);
-    }
-}
-
-impl BuildCache {
     fn cache_version(&self) -> String {
         self.options.version.clone().unwrap_or_default()
     }
-}
-
-fn persistent_codec_registry() -> CodecRegistry {
-    CodecRegistry::new()
-        .with_resolve_record(ResolveRecordCodec::current())
-        .with_module_build_record(ModuleBuildRecordCodec::current())
-        .with_code_generation_record(CodeGenerationRecordCodec::current())
-        .with_asset_render_record(AssetRenderRecordCodec::current())
 }
 
 #[cfg(test)]
@@ -2291,625 +600,4 @@ pub(crate) struct BuildCacheStats {
     pub module_entries: usize,
     pub module_hits: usize,
     pub module_misses: usize,
-}
-
-#[cfg(test)]
-mod tests {
-    use std::{collections::BTreeSet, fs, path::Path};
-
-    use filetime::{FileTime, set_file_mtime};
-    use tempfile::tempdir;
-
-    use super::*;
-    use crate::{ModuleIdentity, snapshot::FileSystemInfo};
-
-    #[derive(Debug, Clone)]
-    struct TestCacheKey(&'static str);
-
-    impl CacheKey for TestCacheKey {
-        fn cache_identifier(&self) -> CacheIdentifier {
-            CacheIdentifier::new(self.0)
-        }
-    }
-
-    #[test]
-    fn filesystem_cache_uses_pinned_idle_timeout_defaults() {
-        let options = CacheOptions::filesystem();
-        assert_eq!(options.idle_timeout, Some(60_000));
-        assert_eq!(options.idle_timeout_for_initial_store, Some(5_000));
-        assert_eq!(options.idle_timeout_after_large_changes, Some(1_000));
-    }
-
-    #[test]
-    fn filesystem_max_age_defaults_to_sixty_days_and_exceeds_u32_milliseconds() {
-        let options = CacheOptions::filesystem();
-        assert_eq!(options.max_age, Duration::from_secs(60 * 24 * 60 * 60));
-        assert!(options.max_age.as_millis() > u128::from(u32::MAX));
-
-        let mut overridden = options;
-        overridden.max_age = Duration::from_millis(u64::from(u32::MAX) + 1);
-        assert_eq!(
-            overridden.max_age,
-            Duration::from_millis(u64::from(u32::MAX) + 1)
-        );
-    }
-
-    #[test]
-    fn shared_persistent_location_warns_without_locking_the_second_writer() {
-        let temp = tempdir().expect("create shared cache location");
-        let mut options = CacheOptions::filesystem();
-        options.cache_location = Some(temp.path().join("cache"));
-        let first = BuildCache::new(options.clone(), SnapshotOptions::default());
-        let second = BuildCache::new(options, SnapshotOptions::default());
-
-        let warnings = second.take_infrastructure_log_events();
-        assert!(warnings.iter().any(|event| {
-            event.level == InfrastructureLogLevel::Warn
-                && event.name == CACHE_WRITER_LOGGER
-                && event.message.contains("another live writer")
-                && event
-                    .message
-                    .contains("without a cross-process lock or merge protocol")
-                && event
-                    .message
-                    .contains("trusted-local,linux-supported,single-writer")
-        }));
-        assert!(first.normal_module_factory().is_enabled());
-        assert!(second.normal_module_factory().is_enabled());
-    }
-
-    #[test]
-    fn cache_facades_scope_identical_identifiers_by_namespace_and_etag() {
-        let build_cache = BuildCache::new(CacheOptions::memory(), SnapshotOptions::default());
-        let code_generation = build_cache.facade::<TestCacheKey, String>(
-            CacheNamespace::new("unpack/code-generation"),
-            CacheItemFamily::CodeGeneration,
-        );
-        let asset_render = build_cache.facade::<TestCacheKey, String>(
-            CacheNamespace::new("unpack/asset-render"),
-            CacheItemFamily::AssetRender,
-        );
-        assert_eq!(
-            code_generation.namespace(),
-            CacheNamespace::new("unpack/code-generation")
-        );
-        assert_eq!(
-            asset_render.namespace(),
-            CacheNamespace::new("unpack/asset-render")
-        );
-        let identifier = TestCacheKey("shared-identifier");
-        let current = CacheETag::new("current");
-        let stale = CacheETag::new("stale");
-
-        code_generation.store(
-            identifier.clone(),
-            Some(current.clone()),
-            "generated source".to_string(),
-        );
-        asset_render.store(
-            identifier.clone(),
-            Some(current.clone()),
-            "rendered asset".to_string(),
-        );
-
-        assert_eq!(
-            code_generation
-                .get(&identifier, Some(&current))
-                .as_deref()
-                .map(String::as_str),
-            Some("generated source")
-        );
-        assert_eq!(
-            asset_render
-                .get(&identifier, Some(&current))
-                .as_deref()
-                .map(String::as_str),
-            Some("rendered asset")
-        );
-        assert!(code_generation.get(&identifier, Some(&stale)).is_none());
-
-        let counters = build_cache.work_counters();
-        assert_eq!(
-            counters.for_family(CacheItemFamily::CodeGeneration),
-            CacheItemWork {
-                hits: 1,
-                misses: 1,
-                stores: 1,
-                restores: 0,
-                evictions: 0,
-            }
-        );
-        assert_eq!(
-            counters.for_family(CacheItemFamily::AssetRender),
-            CacheItemWork {
-                hits: 1,
-                misses: 0,
-                stores: 1,
-                restores: 0,
-                evictions: 0,
-            }
-        );
-    }
-
-    #[test]
-    fn cache_facade_accounts_for_memory_eviction_by_item_family() {
-        let build_cache = BuildCache::new(CacheOptions::memory(), SnapshotOptions::default());
-        let code_generation = build_cache.facade::<TestCacheKey, String>(
-            CacheNamespace::new("unpack/code-generation"),
-            CacheItemFamily::CodeGeneration,
-        );
-        let identifier = TestCacheKey("evicted-identifier");
-
-        code_generation.store(identifier.clone(), None, "generated source".to_string());
-        code_generation.evict_memory(&identifier);
-
-        assert!(code_generation.get(&identifier, None).is_none());
-        assert_eq!(
-            build_cache
-                .work_counters()
-                .for_family(CacheItemFamily::CodeGeneration),
-            CacheItemWork {
-                hits: 0,
-                misses: 1,
-                stores: 1,
-                restores: 0,
-                evictions: 1,
-            }
-        );
-    }
-
-    #[test]
-    fn lower_layer_hit_repopulates_the_earlier_memory_cache() {
-        let build_cache = BuildCache::new(CacheOptions::filesystem(), SnapshotOptions::default());
-        let code_generation = build_cache.facade::<TestCacheKey, String>(
-            CacheNamespace::new("unpack/code-generation"),
-            CacheItemFamily::CodeGeneration,
-        );
-        let identifier = TestCacheKey("restored-identifier");
-
-        code_generation.store(identifier.clone(), None, "generated source".to_string());
-        code_generation.evict_memory(&identifier);
-
-        assert_eq!(
-            code_generation
-                .get(&identifier, None)
-                .as_deref()
-                .map(String::as_str),
-            Some("generated source")
-        );
-        assert_eq!(
-            code_generation
-                .get(&identifier, None)
-                .as_deref()
-                .map(String::as_str),
-            Some("generated source")
-        );
-        assert_eq!(
-            build_cache
-                .work_counters()
-                .for_family(CacheItemFamily::CodeGeneration),
-            CacheItemWork {
-                hits: 2,
-                misses: 0,
-                stores: 1,
-                restores: 1,
-                evictions: 1,
-            }
-        );
-    }
-
-    #[tokio::test]
-    async fn slow_persistent_decode_does_not_block_memory_or_other_persistent_hits()
-    -> std::result::Result<(), Box<dyn std::error::Error>> {
-        let temp = tempdir()?;
-        let resource = temp.path().join("src/dep.js");
-        write(&resource, "export const value = 1;")?;
-        let file_system_info = FileSystemInfo::new();
-        let record = ResolveRecord::new(
-            ModuleIdentity::new(resource.clone()),
-            resource.clone(),
-            BTreeSet::from([resource]),
-            BTreeSet::new(),
-            BTreeSet::new(),
-            &file_system_info,
-            SnapshotStrategy::timestamp(),
-        )
-        .await?;
-        let persistent_key = ResolveRequest::new(temp.path(), "./persistent");
-        let other_persistent_key = ResolveRequest::new(temp.path(), "./other-persistent");
-        let memory_key = ResolveRequest::new(temp.path(), "./memory");
-        let mut options = CacheOptions::filesystem();
-        options.cache_location = Some(temp.path().join("cache"));
-
-        let first = BuildCache::new(options.clone(), SnapshotOptions::default());
-        first
-            .normal_module_factory()
-            .store(persistent_key.clone(), None, record.clone());
-        first
-            .normal_module_factory()
-            .store(other_persistent_key.clone(), None, record.clone());
-        first.flush_to_filesystem()?;
-        drop(first);
-
-        options.readonly = true;
-        let clock = Arc::new(ManualCacheClock::at_millis(1_000));
-        let second =
-            BuildCache::new_with_clock(options, SnapshotOptions::default(), Arc::clone(&clock));
-        let facade = second.normal_module_factory();
-        facade.store(memory_key.clone(), None, record);
-        let restore_entered = Arc::new(std::sync::Barrier::new(2));
-        let restore_release = Arc::new(std::sync::Barrier::new(2));
-        second.install_restore_barrier(restore_entered.clone(), restore_release.clone());
-
-        let restore_facade = facade.clone();
-        let restore_thread = std::thread::spawn(move || restore_facade.get(&persistent_key, None));
-        restore_entered.wait();
-
-        let memory_facade = facade.clone();
-        let (completed_tx, completed_rx) = std::sync::mpsc::channel();
-        let memory_thread = std::thread::spawn(move || {
-            completed_tx
-                .send(memory_facade.get(&memory_key, None).is_some())
-                .expect("memory-hit observation should still be received");
-        });
-        let completed_before_restore_release =
-            completed_rx.recv_timeout(Duration::from_secs(1)).ok();
-
-        let other_persistent_facade = facade.clone();
-        let (other_completed_tx, other_completed_rx) = std::sync::mpsc::channel();
-        let other_persistent_thread = std::thread::spawn(move || {
-            other_completed_tx
-                .send(
-                    other_persistent_facade
-                        .get(&other_persistent_key, None)
-                        .is_some(),
-                )
-                .expect("persistent-hit observation should still be received");
-        });
-        let other_completed_before_restore_release =
-            other_completed_rx.recv_timeout(Duration::from_secs(1)).ok();
-
-        restore_release.wait();
-        assert!(
-            restore_thread
-                .join()
-                .expect("restore thread should finish")
-                .is_some()
-        );
-        memory_thread
-            .join()
-            .expect("memory-hit thread should finish");
-        other_persistent_thread
-            .join()
-            .expect("other persistent-hit thread should finish");
-        assert_eq!(
-            completed_before_restore_release,
-            Some(true),
-            "persistent deserialization must not hold the global BuildCache lock"
-        );
-        assert_eq!(
-            other_completed_before_restore_release,
-            Some(true),
-            "record decoding must not hold the PackFile reader lock"
-        );
-        assert_eq!(
-            clock.calls(),
-            0,
-            "read-only cache hits do not need access stamps"
-        );
-        assert_eq!(
-            second.work_counters().for_family(CacheItemFamily::Resolve),
-            CacheItemWork {
-                hits: 3,
-                misses: 0,
-                stores: 1,
-                restores: 2,
-                evictions: 0,
-            }
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn finite_memory_generations_evict_only_entries_left_unused_for_the_limit() {
-        let mut options = CacheOptions::memory();
-        options.max_memory_generations = Some(1);
-        let build_cache = BuildCache::new(options, SnapshotOptions::default());
-        let code_generation = build_cache.facade::<TestCacheKey, String>(
-            CacheNamespace::new("unpack/code-generation"),
-            CacheItemFamily::CodeGeneration,
-        );
-        let unused = TestCacheKey("unused");
-        let kept = TestCacheKey("kept");
-
-        code_generation.store(unused.clone(), None, "unused source".to_string());
-        code_generation.store(kept.clone(), None, "kept source".to_string());
-        build_cache.on_compilation_completed();
-
-        assert_eq!(
-            code_generation
-                .get(&kept, None)
-                .as_deref()
-                .map(String::as_str),
-            Some("kept source")
-        );
-        build_cache.on_compilation_completed();
-
-        assert!(code_generation.get(&unused, None).is_none());
-        assert_eq!(
-            code_generation
-                .get(&kept, None)
-                .as_deref()
-                .map(String::as_str),
-            Some("kept source")
-        );
-        assert_eq!(
-            build_cache
-                .work_counters()
-                .for_family(CacheItemFamily::CodeGeneration)
-                .evictions,
-            1
-        );
-    }
-
-    #[test]
-    fn finite_memory_generations_keep_entries_until_the_completed_generation_boundary() {
-        let mut options = CacheOptions::memory();
-        options.max_memory_generations = Some(2);
-        let build_cache = BuildCache::new(options, SnapshotOptions::default());
-        let code_generation = build_cache.facade::<TestCacheKey, String>(
-            CacheNamespace::new("unpack/code-generation"),
-            CacheItemFamily::CodeGeneration,
-        );
-        let identifier = TestCacheKey("generation-boundary");
-
-        code_generation.store(identifier.clone(), None, "source".to_string());
-        build_cache.on_compilation_completed();
-        build_cache.on_compilation_completed();
-
-        assert_eq!(
-            code_generation
-                .get(&identifier, None)
-                .as_deref()
-                .map(String::as_str),
-            Some("source")
-        );
-        build_cache.on_compilation_completed();
-        build_cache.on_compilation_completed();
-        build_cache.on_compilation_completed();
-
-        assert!(code_generation.get(&identifier, None).is_none());
-        assert_eq!(
-            build_cache
-                .work_counters()
-                .for_family(CacheItemFamily::CodeGeneration)
-                .evictions,
-            1
-        );
-    }
-
-    #[test]
-    fn etag_mismatch_does_not_refresh_an_entrys_generation() {
-        let mut options = CacheOptions::memory();
-        options.max_memory_generations = Some(2);
-        let build_cache = BuildCache::new(options, SnapshotOptions::default());
-        let code_generation = build_cache.facade::<TestCacheKey, String>(
-            CacheNamespace::new("unpack/code-generation"),
-            CacheItemFamily::CodeGeneration,
-        );
-        let identifier = TestCacheKey("etag-mismatch");
-
-        code_generation.store(
-            identifier.clone(),
-            Some(CacheETag::new("expected")),
-            "source".to_string(),
-        );
-        build_cache.on_compilation_completed();
-        assert!(
-            code_generation
-                .get(&identifier, Some(&CacheETag::new("different")))
-                .is_none()
-        );
-        build_cache.on_compilation_completed();
-        build_cache.on_compilation_completed();
-
-        assert!(
-            code_generation
-                .get(&identifier, Some(&CacheETag::new("expected")))
-                .is_none()
-        );
-        assert_eq!(
-            build_cache
-                .work_counters()
-                .for_family(CacheItemFamily::CodeGeneration)
-                .evictions,
-            1
-        );
-    }
-
-    #[test]
-    fn unbounded_memory_generations_never_age_entries() {
-        let build_cache = BuildCache::new(CacheOptions::memory(), SnapshotOptions::default());
-        let code_generation = build_cache.facade::<TestCacheKey, String>(
-            CacheNamespace::new("unpack/code-generation"),
-            CacheItemFamily::CodeGeneration,
-        );
-        let identifier = TestCacheKey("unbounded");
-
-        code_generation.store(identifier.clone(), None, "source".to_string());
-        for _ in 0..100 {
-            build_cache.on_compilation_completed();
-        }
-
-        assert_eq!(
-            code_generation
-                .get(&identifier, None)
-                .as_deref()
-                .map(String::as_str),
-            Some("source")
-        );
-        assert_eq!(
-            build_cache
-                .work_counters()
-                .for_family(CacheItemFamily::CodeGeneration)
-                .evictions,
-            0
-        );
-    }
-
-    #[test]
-    fn zero_filesystem_memory_generations_keep_no_memory_layer() {
-        let mut options = CacheOptions::filesystem();
-        options.max_memory_generations = Some(0);
-        let build_cache = BuildCache::new(options, SnapshotOptions::default());
-        let code_generation = build_cache.facade::<TestCacheKey, String>(
-            CacheNamespace::new("unpack/code-generation"),
-            CacheItemFamily::CodeGeneration,
-        );
-        let identifier = TestCacheKey("persistent-only");
-
-        code_generation.store(identifier.clone(), None, "source".to_string());
-
-        assert_eq!(
-            build_cache
-                .inner
-                .cache
-                .lock()
-                .expect("build cache data mutex should not be poisoned")
-                .entry_count(CacheItemFamily::CodeGeneration),
-            0
-        );
-        assert_eq!(
-            code_generation
-                .get(&identifier, None)
-                .as_deref()
-                .map(String::as_str),
-            Some("source")
-        );
-        assert_eq!(
-            build_cache
-                .work_counters()
-                .for_family(CacheItemFamily::CodeGeneration),
-            CacheItemWork {
-                hits: 1,
-                misses: 0,
-                stores: 1,
-                restores: 0,
-                evictions: 0,
-            }
-        );
-    }
-
-    #[tokio::test]
-    async fn resolve_record_context_snapshot_invalidates_directory_entry_changes()
-    -> std::result::Result<(), Box<dyn std::error::Error>> {
-        let temp = tempdir()?;
-        let context = temp.path().join("src");
-        let resource = context.join("dep.js");
-        write(&resource, "export const value = 'js';")?;
-        let original_mtime = FileTime::from_system_time(fs::metadata(&context)?.modified()?);
-        let file_system_info = FileSystemInfo::new();
-        let record = ResolveRecord::new(
-            ModuleIdentity::new(resource.clone()),
-            resource,
-            BTreeSet::new(),
-            BTreeSet::from([context.clone()]),
-            BTreeSet::new(),
-            &file_system_info,
-            SnapshotStrategy::timestamp(),
-        )
-        .await?;
-
-        write(context.join("dep.ts"), "export const value = 'ts';")?;
-        set_file_mtime(&context, original_mtime)?;
-
-        assert!(
-            !record
-                .is_valid(&file_system_info, SnapshotStrategy::timestamp())
-                .await
-        );
-
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn memory_hits_refresh_persistent_access_before_max_age_gc()
-    -> std::result::Result<(), Box<dyn std::error::Error>> {
-        let temp = tempdir()?;
-        let resource = temp.path().join("src/dep.js");
-        write(&resource, "export const value = 1;")?;
-        let file_system_info = FileSystemInfo::new();
-        let record = ResolveRecord::new(
-            ModuleIdentity::new(resource.clone()),
-            resource.clone(),
-            BTreeSet::from([resource]),
-            BTreeSet::new(),
-            BTreeSet::new(),
-            &file_system_info,
-            SnapshotStrategy::timestamp(),
-        )
-        .await?;
-        let first_key = ResolveRequest::new(temp.path(), "./first");
-        let recent_key = ResolveRequest::new(temp.path(), "./recent");
-        let clock = Arc::new(ManualCacheClock::at_millis(1_000));
-        let mut options = CacheOptions::filesystem();
-        options.cache_location = Some(temp.path().join("cache"));
-        options.max_age = Duration::from_millis(100);
-
-        let first =
-            BuildCache::new_with_clock(options.clone(), SnapshotOptions::default(), clock.clone());
-        first
-            .normal_module_factory()
-            .store(first_key.clone(), None, record.clone());
-        first.flush_to_filesystem()?;
-        assert_eq!(pack_revision(&options), 1);
-
-        clock.set_millis(1_050);
-        let second =
-            BuildCache::new_with_clock(options.clone(), SnapshotOptions::default(), clock.clone());
-        let facade = second.normal_module_factory();
-        assert!(facade.get(&first_key, None).is_some());
-        second.flush_to_filesystem()?;
-        assert_eq!(pack_revision(&options), 2);
-
-        clock.set_millis(1_120);
-        assert!(facade.get(&first_key, None).is_some());
-        second.flush_to_filesystem()?;
-        assert_eq!(pack_revision(&options), 3);
-
-        clock.set_millis(1_221);
-        facade.store(recent_key.clone(), None, record.clone());
-        second.flush_to_filesystem()?;
-        assert_eq!(pack_revision(&options), 4);
-
-        clock.set_millis(1_222);
-        facade.store(recent_key.clone(), None, record);
-        second.flush_to_filesystem()?;
-        assert_eq!(pack_revision(&options), 5);
-
-        let third = BuildCache::new_with_clock(options, SnapshotOptions::default(), clock);
-        let facade = third.normal_module_factory();
-        assert!(facade.get(&first_key, None).is_none());
-        assert!(facade.get(&recent_key, None).is_some());
-        Ok(())
-    }
-
-    fn pack_revision(options: &CacheOptions) -> u64 {
-        PackFile::open(
-            options
-                .cache_location
-                .as_ref()
-                .expect("filesystem cache should have a location"),
-            persistent_codec_registry(),
-        )
-        .revision()
-    }
-
-    fn write(path: impl AsRef<Path>, source: &str) -> io::Result<()> {
-        let path = path.as_ref();
-        if let Some(parent) = path.parent() {
-            fs::create_dir_all(parent)?;
-        }
-        fs::write(path, source)
-    }
 }
